+++ src/Root.tsx
@@ -6,19 +6,17 @@ import { Home } from './pages/Home';
 import { People } from './pages/People';
 import { NotFound } from './pages/NotFound';
 
-export const Root = () => {
-  return (
-    <HashRouter>
-      <Routes>
-        <Route path="/" element={<App />}>
-          <Route index element={<Home />} />
-          <Route path="home" element={<Navigate to=".." />} />
-          <Route path="people">
-            <Route path=":slug?" element={<People />} />
-          </Route>
-          <Route path="*" element={<NotFound />} />
+export const Root = () => (
+  <HashRouter>
+    <Routes>
+      <Route path="/" element={<App />}>
+        <Route index element={<Home />} />
+        <Route path="home" element={<Navigate to=".." />} />
+        <Route path="people">
+          <Route path=":slug?" element={<People />} />
         </Route>
-      </Routes>
-    </HashRouter>
-  );
-};
+        <Route path="*" element={<NotFound />} />
+      </Route>
+    </Routes>
+  </HashRouter>
+);
