+++ src/components/navigation/Navigation.tsx
@@ -1,0 +1,25 @@
+import classNames from 'classnames';
+import { NavLink } from 'react-router-dom';
+
+const getLinkClass = ({ isActive }: { isActive: boolean }) => classNames(
+  'navbar-item',
+  { 'has-background-grey-lighter': isActive },
+);
+
+export const Navigation = () => (
+  <div className="navbar-brand">
+    <NavLink
+      to="/"
+      className={getLinkClass}
+    >
+      Home
+    </NavLink>
+
+    <NavLink
+      to="/people"
+      className={getLinkClass}
+    >
+      People
+    </NavLink>
+  </div>
+);
