+++ src/pages/People.tsx
@@ -46,7 +46,6 @@ export const People = () => {
           {!!peopleList.length && !isLoading && (
             <PeopleTable />
           )}
-
         </div>
       </div>
     </>
